homelab/voice/requirements.txt
paho-mqtt>=2.0.0
httpx>=0.25.0
